rag_gemini/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
python-dotenv
psycopg2-binary
asyncpg
pgvector
tavily-python
mem0ai
langchain
langchain-google-genai
langchain-community
langchain-postgres
pypdf
openpyxl
docx2txt
pandas
